Move worker downloading to MultipleLifetime worker
Worker script needs to be downloaded from packager and patched only once. However it was downloaded every time when new SingleLifetime worker was starting. Moving the download logic to MultipleLifetime worker would improve reload time.

Also extract worker downloading logic for better testability
